% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_detection.R
\name{accuracy_d}
\alias{accuracy_d}
\title{Determine accuracy d'}
\usage{
accuracy_d(option.tab, loglinear = TRUE, aggregateBy = c("id"),
  colnames = c("hit_by_option", "false_alarm_by_option"))
}
\arguments{
\item{option.tab}{A long table containing DOMC response data by
response option (created by `get_item_table(response_data, IDs,
by_option = TRUE)`)}

\item{loglinear}{Boolean; should a loglinear correction be applied
(defaults to `TRUE`)}

\item{aggregateBy}{Must be `c("option", "id")`, `"option"`, or
`"id"`. Determines what variables the yes-rate is aggregated
over; the default is `c("id")` in which case d' is computed for
each test-taker across all decisions.}

\item{colnames}{Indicates the names of the columns containing the
hits and false alarms in `long.tab`}
}
\value{
A data.frame in long format where each row contains a measure
    of the accuracy parameter.
}
\description{
Determine accuracy d'
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
