% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_criterion.R
\name{ci}
\alias{ci}
\title{Compute the half confidence interval of a vector}
\usage{
ci(ci, x, na.rm = FALSE)
}
\arguments{
\item{ci}{the confidence level as a natural number (e.g., 95)}

\item{x}{a vector of values}

\item{na.rm}{Boolean. Remove NA values?}
}
\description{
Compute the half confidence interval of a vector
}
