% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_criterion.R
\name{ci95}
\alias{ci95}
\title{Compute the half of the 95\% confidence interval of a vector}
\usage{
ci95(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of values}

\item{na.rm}{Boolean. Remove NA values?}
}
\description{
Compute the half of the 95\% confidence interval of a vector
}
