% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Settings to collect Dota2 matches.}
\usage{
config(key = NULL, game_mode = c(2, 22), lobby_type = 7, skill = 3,
  duration = 900, public_account_id = 5, n_history_matches = 10)
}
\arguments{
\item{key}{The api key obtained from Steam. If you don't have one please visit
\url{https://steamcommunity.com/dev} in order to do so.}

\item{game_mode}{The game mode can be a vector with one or more of these values:
\itemize{
  \item 0 - Unknown
  \item 1 - All Pick
  \item 2 - Captain's Mode
  \item 3 - Random Draft
  \item 4 - Single Draft
  \item 5 - All Random
  \item 6 - Intro
  \item 7 - Diretide
  \item 8 - Reverse Captain's Mode
  \item 9 - The Greeviling
  \item 10 - Tutorial
  \item 11 - Mid Only
  \item 12 - Least Played
  \item 13 - Limited Heroes
  \item 14 - Compendium Matchmaking
  \item 15 - Custom
  \item 16 - Captain's Draft
  \item 17 - Balanced Draft
  \item 18 - Ability Draft
  \item 19 - Event
  \item 20 - Random Death Match
  \item 21 - 1v1 Mid
  \item 22 - All Draft
  \item 23 - Turbo
  \item 24 - Mutation
}

(Default: 2, 22)}

\item{lobby_type}{The lobby type can be a vector with one or more of these values:
\itemize{
  \item 0 - Normal
  \item 1 - Practice
  \item 2 - Tournament
  \item 3 - Tutorial
  \item 4 - Coop Bots
  \item 5 - Ranked Team MM
  \item 6 - Ranked Solo MM
  \item 7 - Ranked
  \item 8 - 1v1 Mid
  \item 9 - Battle Cup
}

(Default: 7)}

\item{skill}{The skill bracket can be a vector with one or more of these values (except 0):
\itemize{
  \item 0 - Any (If is set with this value, just it will be used)
  \item 1 - Normal
  \item 2 - High
  \item 3 - Very High
}

(Default: 3)}

\item{duration}{The minimum duration of a match in seconds. (Default: 900)}

\item{public_account_id}{The minimum of public account's id. (Default: 5)}

\item{n_history_matches}{The amount of history matches that must be collected for each player in the
game. (Default: 10)}

\item{database_name}{The name of database that the data will be stored in MongoDB. (Default: dota2)}
}
\value{
A message saying that the settings have been logged.
}
\description{
This function is used to define the settings that will be used to collect the data,
    such as type of game, skill etc ...
}
\examples{

# An available API key
config(key = 'xxxxx-xxxxx')
}
