% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_detection.R
\name{criterion_c}
\alias{criterion_c}
\title{Determine response criterion c by option position per test-taker}
\usage{
criterion_c(option.tab, loglinear = TRUE, aggregateBy = c("option_position",
  "id"), colnames = c("hit_by_option", "false_alarm_by_option"))
}
\arguments{
\item{loglinear}{Boolean; should a loglinear correction be applied
(defaults to TRUE which is *strongly* recommended to compute
response criteria)}

\item{aggregateBy}{Must be `c("option_position", "id")`,
`"option_position"`, or `"id"`. Determines what variables the
yes-rate is aggregated over; the default is `c("option_position",
"id")` in which case the response criterion is computed for each
test-taker by option.}

\item{colnames}{What are the columns containing hit and false alarm
scores called? Generally the default should not be changed.}

\item{long.tab}{A long table containing DOMC response data by
response option (created by `get_item_table(responseDataSample,
IDs, by_option = TRUE)`)}
}
\description{
Determine response criterion c by option position per test-taker
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
