% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_criterion.R
\name{error_bar}
\alias{error_bar}
\title{Draw an error bar to a plot}
\usage{
error_bar(x, y, upper, lower = upper, length = 0.1, width = 1.5, col)
}
\arguments{
\item{x}{The x coordinates}

\item{y}{The y coordinates}

\item{upper}{The upper bound of the error interval (in relation to
the y-coordinates)}

\item{lower}{The lower bound of the error interval (in relation to
the y-coordinates)}

\item{length}{the width of the bar end}

\item{width}{the width of the error bar line}

\item{col}{The color of the error bar}
}
\description{
Draw an error bar to a plot
}
