% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_morbist.R
\name{estimateAccuracy}
\alias{estimateAccuracy}
\title{Estimate accuracy for a given solving probability}
\usage{
estimateAccuracy(probability, n_options = NULL, sd_sol = 1, type,
  criterionVector = NULL, item = NULL)
}
\arguments{
\item{probability}{probability to solve the item correctly (must not
be 1 or 0)}

\item{n_options}{How many options does the item have}

\item{sd_sol}{Standard deviation of the solution
distribution. Distractor always has SD of 1}

\item{type}{either "mc" or "domc"}
}
\value{
The estimated accuracy
}
\description{
This function is the inverse to \code{mc_analytical} and
\code{domc_analytical}: It determines the accuracy parameter that
corresponds to a given solving probability. Can be used to infer
appropriate accuracy values for empirical data.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
