% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heroes_details.R
\name{get_heroes_details}
\alias{get_heroes_details}
\title{Details about each hero}
\usage{
get_heroes_details(key)
}
\arguments{
\item{key}{The api key obtained from Steam. If you don't have one please visit
\url{https://steamcommunity.com/dev} in order to do so.}
}
\value{
A data frame where each row contains a hero and his/her attributes.
}
\description{
Details about each hero such as STR, AGI, INT, Range of day vision or night etc...
}
\examples{
df <- get_hero_detais(key = 'xxxxx-xxxxx')
head(df)

}
