% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_details_2.R
\name{get_match_details_2}
\alias{get_match_details_2}
\title{Get the information about a specific match}
\usage{
get_match_details_2(...)
}
\description{
This function is a wrapper of RDota2::get_match_details. The purpose of this
    function is to try to collect the details matches until your success or fail, without
    returning a error.
}
