% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_history_2.R
\name{get_match_history_2}
\alias{get_match_history_2}
\title{A list of matches}
\usage{
get_match_history_2(...)
}
\description{
This function is a wrapper of RDota2::get_match_history. The purpose of this
    function is to try to collect the history of matches until your success or fail, without
    returning a error.
}
