% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_from_unipark.R
\name{get_response_data_person}
\alias{get_response_data_person}
\title{Convert DOMC response data from Unipark output to R list}
\usage{
get_response_data_person(JsonString, itemIDs = NULL, itemPrefix = "item")
}
\arguments{
\item{JsonString}{The JSON string, usually stored in the column of an
Unipark-exported csv}

\item{itemIDs}{A numeric vector corresponding to the IDs of all items
to be exported. Change only if only a subset of items is to be
returned; otherwise do not change the default value `NULL`.}

\item{itemPrefix}{By default the function assumes that each item is
represented in the JSON string by the name "itemx" where x is the
ID of the item. If the prefix is anything other than "item", this
argument can be used to tell the function this prefix.}
}
\value{
A \code{list} containing the test data for one test taker
  \item{item1 ... itemx}{Item data for the x test items}
}
\description{
Converts response data for one test-taker from Unipark JSON String to
a `list`.
}
\details{
Each entry of the returned \code{list} contains data
    detailing a test-taker's response to one item.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
