% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_morbist.R
\name{mc_analytical}
\alias{mc_analytical}
\title{Probability that a MC item is solved correctly given accuracy.}
\usage{
mc_analytical(accuracy, n_options = 4, sd_sol = 1)
}
\arguments{
\item{accuracy}{Accuracy value of an assumed test taker}

\item{n_options}{number of answer options in the MC item}

\item{sd_sol}{Standard deviation of the solution
distribution. Distractor SD is always 1}
}
\value{
The expected probability that the item is solved correctly
}
\description{
This function can be used to analytically determine the expected
probability that an MC item is solved correctly, as opposed to the
simulation approach used in \code{workMCItem}.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
