% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_criterion.R
\name{plot_criterion}
\alias{plot_criterion}
\title{Plot the response criterion by option position, possibly using a
grouping variable (can also be used to plot any other values in
dependence of option position and acceptance reluctance)}
\usage{
plot_criterion(tab, ylim = c(-1, 1), xlim = c(1 - 0.2, nrow(tab) + 0.2),
  txt.cords = list(c(1, 0.3), c(2.8, 0.35)), col = c("black", "red"),
  xlab = "Position of answer", ylab = "Response criterion c",
  labelCriterion = TRUE, add = FALSE, pch = c(16, 15), lty = 1,
  lwd = 1.3, plotNeutral = TRUE, cex = 1.4, err.bar = NULL,
  type = "b", txt = colnames(tab), labelx = TRUE, dev = 0.07,
  main = "", lwd_arrow = 1, ...)
}
\arguments{
\item{tab}{A data table with rows = criterion by position (row 1:
criterion for option 1) and columns distinguishing different
levels of grouping factor}

\item{ylim}{Limits of y-axis}

\item{txt.cords}{A `list`, containing two vectors of x and y
coordinates specifying where the factor level descriptions are
printed}

\item{col}{A vector of two colors specifying the coloring of two
levels of the factor}

\item{xlab}{Label of the x-axis}

\item{ylab}{Label of the y-axis}

\item{labelCriterion}{boolean value, if `TRUE`, on the right axis the
response criterion is labeled as "conservative", "neutral", and
"liberal"}

\item{add}{Boolean. Should the data be drawn into an existing plot?}

\item{pch}{The type of the data points}

\item{lty}{the type of the lines drawn}

\item{lwd}{the width of the lines drawn}

\item{plotNeutral}{Boolean. Should a horizontal line at y=0 be drawn
to illustrate a neutral response criterion?}

\item{cex}{the size of the dots}

\item{err.bar}{If the error bars should be plotted, pass a table of
the same dimension as `tab`. Each entry is the length of the
error bar to each side of the point (i.e.: the value should be
half of the width of the confidence interval if a confidence
interval is to be plotted)}

\item{type}{The type of plot, see `?plot`}

\item{txt}{A vector of text describing the levels of the grouping
variable}

\item{labelx}{Vector to label the ticks on the x-axis}

\item{dev}{A single value indicating the margin by which data points
on the same x-axis point should be separated to better illustrate
different grouping points}

\item{main}{The caption of the plot}

\item{lwd_arrow}{The thickness of the arrow illustrating liberal
vs. conservative responses}

\item{...}{further arguments passed to `plot`}
}
\description{
Plot the response criterion by option position, possibly using a
grouping variable (can also be used to plot any other values in
dependence of option position and acceptance reluctance)
}
