% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_morbist.R
\name{plot_morbist}
\alias{plot_morbist}
\title{Plot the parameters characterizing the MORBIST model}
\usage{
plot_morbist(x, y, showCriterion = TRUE, labelCriterion = FALSE,
  distCritLabel = 0.425, cex.axis = 1, cex.label = 1, cex.crit = 1,
  cex.acc = 1, caption = "evidence strength", solutions = "solutions",
  distractors = "distractors", distLabelDistributions = 2.3,
  showAccuracy = TRUE, distAccLabel = 0.46)
}
\arguments{
\item{x}{An accuracy value (d')}

\item{y}{The response criterion vector}

\item{showCriterion}{Boolean. Should the criterion location be drawn?}

\item{labelCriterion}{label response criterion with *c*; only has an
effect if showCriterion is TRUE}

\item{distCritLabel}{If the criterion label is displayed: how far top
should it be (higher values move the label further away from the
plot)}

\item{cex.axis}{magnification of the decision variable axis (x-axis)
values}

\item{cex.label}{controls magnification of text size of x-axis label
and "distractors" and "solutions" text}

\item{cex.acc}{controls magnification the accuracy label}

\item{caption}{label for decision variable axis (x-axis)}

\item{solutions}{caption for solution distribution}

\item{distractors}{caption for distractor distribution}

\item{distLabelDistributions}{how away from the distribution plots
should the labels for the distributions be (higher values move
the label further away from the plot)}

\item{showAccuracy}{Boolean - illustrate accuracy d'?}

\item{distAccLabel}{If the accuracy value is displayed: how far top
should it be (higher values move the label further away from the
plot)}

\item{cex.label}{controls magnification the text of the criterion
labels}
}
\value{
None
}
\description{
Plots decision variable distribution of targets and and lures
(assuming a normal distribution of evidence) for a test-taker; can
also illustrate a response criterion vector
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
