% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_chi2}
\alias{print_chi2}
\title{Printing the results of a significance test for a correlation coefficient}
\usage{
print_chi2(table = NULL, es = TRUE, chi2.object = NULL, decimals = 2,
  decimals_p = 3)
}
\arguments{
\item{table}{a contingency table that is passed to `chisq.test`}

\item{es}{Boolean. Should the phi coefficient be printed (only
advised for 2x2 contingency tables!)}

\item{chi2.object}{an object that is returned by `chisq.test`}

\item{decimals}{how many decimals should be printed}

\item{decimals_p}{how many decimals should be printed for the p-value
(defaults to 3)}
}
\value{
A string describing the chi-squared test
}
\description{
Printing the results of a significance test for a correlation coefficient
}
