% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_ttest}
\alias{print_ttest}
\title{function for printing results of t-test}
\usage{
print_ttest(t_object, d_object, decimals = 2, decimals_p = 3)
}
\description{
function for printing results of t-test
}
