% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_criterion.R
\name{se}
\alias{se}
\title{Compute standard error}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of values}

\item{na.rm}{Boolean. Remove NA values?}
}
\description{
Compute standard error
}
