% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_morbist.R
\name{simulate_morbist}
\alias{simulate_morbist}
\title{Simulate test responses in the MORBIST model}
\usage{
simulate_morbist(item_number, option_number, n_respondents, type,
  avg_acc = 1.5, sd_acc = 0.5, accuracies = NULL, criterion_list,
  sd_sol = 1, sd_dis = 1)
}
\arguments{
\item{item_number}{How many items are simulated for each test-taker}

\item{option_number}{How many response options should each item have}

\item{n_respondents}{How many test-takers are to be simulated}

\item{type}{Which type of test is be simulated? Must be "mc" or
"domc".}

\item{avg_acc}{Mean accuracy (d') of the test-takers}

\item{sd_acc}{Standard deviation of the test-takers' accuracy values
d'.}

\item{accuracies}{If not NULL, it must be a vector of length
`n_respondents` containing an accuracy value d' for each test
taker. If this argument is passed, it overrides the values of
avg_acc and sd_acc. Defaults to NULL.}

\item{criterion_list}{A list of criterion vectors of length
`testTaker`.}

\item{sd_sol}{standard deviation of solution distribution}

\item{sd_dis}{standard deviation of distractor distribution}
}
\value{
a \code{list} containing testData lists as returned by
\code{\link{workTest}} and test taker accuracy and bias vector
    \item{data}{\code{list}} of testData returned by \code{\link{workTest}}
    \item{accuracy}{\code{list} of test taker accuracy values}
    \item{bias}{\code{list} of test taker response criterion vectors}
}
\description{
Simulates a specified number of test-takers working on a specified
number of items in MC or DOMC format. The accuracy of test-takers can
be generated randomly or is passed explicitly.
}
\examples{

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
