% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOTA2-package.R
\docType{package}
\name{DOTA2-package}
\alias{DOTA2-package}
\title{The DOTA2 Package}
\description{
Discrete option test analysis.
}
\details{
\tabular{ll}{
Package: \tab DOTA2\cr
Type: \tab Package\cr
Version: \tab 0.0.03\cr
Date: \tab 2018-02-15\cr
Depends: \tab R (>= 3.0.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

Analyze and simulate responses in sequential DOMC tests
}
\author{
Martin Papenberg <martin.papenberg@hhu.de>
}
\keyword{package}
